// src/maintenance/domain/enums/service-type.enum.ts
export enum ServiceType {
  OIL_CHANGE = 'OIL_CHANGE',
  TIRE_REPLACEMENT = 'TIRE_REPLACEMENT',
  BRAKE_SERVICE = 'BRAKE_SERVICE',
  CHAIN_MAINTENANCE = 'CHAIN_MAINTENANCE',
  COOLANT_CHANGE = 'COOLANT_CHANGE',
  SPARK_PLUG_REPLACEMENT = 'SPARK_PLUG_REPLACEMENT',
  BATTERY_REPLACEMENT = 'BATTERY_REPLACEMENT',
  AIR_FILTER_REPLACEMENT = 'AIR_FILTER_REPLACEMENT',
  VALVE_ADJUSTMENT = 'VALVE_ADJUSTMENT',
  GENERAL_INSPECTION = 'GENERAL_INSPECTION',
  CUSTOM = 'CUSTOM',
}